refactor: move generate_key to it's own binary
