+++ router/Dockerfile
@@ -26,7 +26,8 @@ RUN --mount=type=cache,target=/usr/local/cargo/registry \
 RUN --mount=type=cache,target=/usr/local/cargo/registry \
     --mount=type=cache,target=/app/target \
     cargo build --release -p gas-killer-router && \
-    cp /app/target/release/gas-killer-router /usr/local/bin/gas-killer-router
+    cp /app/target/release/gas-killer-router /usr/local/bin/gas-killer-router && \
+    cp /app/target/release/generate_key /usr/local/bin/generate_key
 
 # Runtime stage
 FROM debian:bookworm-slim
@@ -42,8 +43,9 @@ RUN apt-get update && apt-get install -y \
 # Create a non-root user
 RUN useradd -m -u 1000 -s /bin/bash appuser
 
-# Copy the binary from builder
+# Copy binaries from builder
 COPY --from=builder /usr/local/bin/gas-killer-router /usr/local/bin/gas-killer-router
+COPY --from=builder /usr/local/bin/generate_key /usr/local/bin/generate_key
 
 # Copy configuration files
 COPY config /app/config
